refactor: Move GraphQL schema management into crate
* Move all definitions of GraphQL objects, loaders, and schemas to a new
  crate `graphql-api`.
* Make `game_api` crate to depend on this new crate.
* Create new `graphql-schema-generator` binary that generates the
  GraphQL SDL and outputs it to the provided file, based on command line
  arguments. This binary is based on the new `graphql-api` crate.
* Move the `UnknownRole` and `InvalidMappackId` error variants from the
  `game_api` crate to the `records-lib` crate.
* Rename the `HasExpireTime` trait from the `game_api` crate to
  `Expirable` and move it to the `records-lib` crate.
* Move each GraphQL object definition to their own module for better
  modularity.
