+++ crates/entity/src/entities/checkpoint_times.rs
@@ -1,21 +1,17 @@
-use async_graphql::SimpleObject;
 use sea_orm::entity::prelude::*;
 
 /// Represents the time on a checkpoint on a map associated to a record.
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, SimpleObject)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "checkpoint_times")]
-#[graphql(name = "CheckpointTime")]
 pub struct Model {
     /// The checkpoint number. It starts at 0.
     #[sea_orm(primary_key, auto_increment = false)]
     pub cp_num: u32,
     /// The ID of the related map.
     #[sea_orm(primary_key, auto_increment = false)]
-    #[graphql(skip)]
     pub map_id: u32,
     /// The ID of the related record.
     #[sea_orm(primary_key, auto_increment = false)]
-    #[graphql(skip)]
     pub record_id: u32,
     /// The time in milliseconds on this checkpoint.
     pub time: i32,
