+++ crates/graphql-api/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "graphql-api"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+async-graphql = { workspace = true }
+entity = { path = "../entity" }
+records-lib = { path = "../records_lib" }
+sea-orm = { workspace = true }
+reqwest = { workspace = true }
+serde = { workspace = true }
+tokio = { workspace = true }
+deadpool-redis = { workspace = true }
+chrono = { workspace = true }
+futures = { workspace = true }
