+++ crates/graphql-api/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod loaders;
+pub mod objects;
+pub mod schema;
