+++ crates/graphql-api/src/loaders/mod.rs
@@ -1,0 +1,4 @@
+pub mod event;
+pub mod event_category;
+pub mod map;
+pub mod player;
