+++ crates/graphql-api/src/objects/mod.rs
@@ -1,0 +1,24 @@
+pub mod checkpoint_time;
+pub mod medal_times;
+pub mod ranked_record;
+pub mod root;
+pub mod sort_state;
+
+pub mod event;
+pub mod event_category;
+pub mod event_edition;
+pub mod event_edition_map;
+pub mod event_edition_map_ext;
+pub mod event_edition_player;
+pub mod event_edition_player_categorized_rank;
+pub mod event_edition_player_rank;
+
+pub mod mappack;
+pub mod mappack_map;
+pub mod mappack_player;
+
+pub mod map;
+pub mod player_rating;
+pub mod related_edition;
+
+pub mod player;
