+++ crates/graphql-api/src/objects/sort_state.rs
@@ -1,0 +1,7 @@
+use async_graphql::Enum;
+
+#[derive(Clone, Copy, PartialEq, PartialOrd, Eq, Ord, Enum)]
+pub(crate) enum SortState {
+    Sort,
+    Reverse,
+}
