+++ crates/graphql-schema-generator/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "graphql-schema-generator"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow.workspace = true
+clap = { version = "4.5.48", features = ["derive"] }
+graphql-api = { path = "../graphql-api" }
